KPMP-51: Break out attach files into separate component.  Disable attach button if no files added
